+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export { default as WeekView } from './WeekView';
 export { default as MonthView } from './MonthView';
 export { default as EventFormView } from './EventFormView';
+export { default as EventListView } from './EventListView';
